Added Python 3 compatibility; cleaned up date string returns.  Fixed single url query.

+++ gdelt/base.py
@@ -102,8 +102,10 @@ class gdelt(object):
 
 
         if int(self.version) == 1:
+
             if self.table is "mentions":
-                raise ('GDELT 1.0 does not have the "mentions" table. Use "events" or "gkg".')
+                raise BaseException('GDELT 1.0 does not have the "mentions" table. Specify the "events" or'
+                                    ' "gkg" table.')
             else:
                 pass
 
@@ -112,12 +114,15 @@ class gdelt(object):
 
             if self.table == 'events':
 
+
                 if self.coverage is True:
+
                     self.download_list = (urlsv1events(v1RangerCoverage(dateRanger(self.date))))
 
                 # print ("1 events", urlsv1events(self.datesString))
                 # print (urlsv2events(v2RangerNoCoverage(dateRanger(self.date))))
                 else:
+                    print("I'm here at line 125")
                     self.download_list = (urlsv1events(v1RangerNoCoverage(dateRanger(self.date))))
 
         #####################################
@@ -128,7 +133,7 @@ class gdelt(object):
             if self.table == 'events':
                 columns = self.events_columns
                 if self.coverage is True:
-                    print('in coverage')
+
                     self.download_list = (urlsv2events(v2RangerCoverage(dateRanger(self.date))))
                 else:
                     self.download_list = (urlsv2events(v2RangerNoCoverage(dateRanger(self.date))))
@@ -136,7 +141,7 @@ class gdelt(object):
             if self.table == 'gkg':
                 columns = self.gkg_columns
                 if self.coverage is True:
-                    print('in coverage')
+
                     self.download_list = (urlsv2gkg(v2RangerCoverage(dateRanger(self.date))))
                 else:
                     self.download_list = (urlsv2gkg(v2RangerNoCoverage(dateRanger(self.date))))
@@ -145,30 +150,51 @@ class gdelt(object):
             if self.table == 'mentions':
                 columns = self.mentions_columns
                 if self.coverage is True:
-                    print('in coverage')
+
                     self.download_list = (urlsv2mentions(v2RangerCoverage(dateRanger(self.date))))
+
                 else:
+
                     self.download_list = (urlsv2mentions(v2RangerNoCoverage(dateRanger(self.date))))
-
-                    # print ("2 mentions", urlsv2mentions(self.datesString))
-
-        pool = Pool(processes=cpu_count())
-
-        downloaded_dfs = list(pool.imap_unordered(mp_worker, self.download_list))
-        pool.close()
-        pool.terminate()
-        pool.join()
-        results = pd.concat(downloaded_dfs)
-        results.reset_index(drop=True, inplace=True)
-
-        results.columns = columns
-
-        self.final = results
 
         #########################
         # DEBUG Print
         #########################
-        # print (self.version, self.table, self.coverage, self.datesString)
+        # print (self.version, self.table, self.coverage, self.datesString,self.download_list)
+
+        #########################
+        # Download section
+        #########################
+
+
+
+        if isinstance(self.datesString, str):
+
+            results = mp_worker(self.download_list)
+
+
+
+        else:
+
+            pool = Pool(processes=cpu_count())
+            downloaded_dfs = list(pool.imap_unordered(mp_worker, self.download_list))
+            pool.close()
+            pool.terminate()
+            pool.join()
+            results = pd.concat(downloaded_dfs)
+            del downloaded_dfs
+            results.reset_index(drop=True, inplace=True)
+
+
+
+
+        results.columns = columns
+
+        if (len(results)) == 0:
+            raise ValueError("This GDELT query returned no data. Check internet connection or query parameters"
+                             " and retry")
+
+        self.final = results
 
         #########################
         # Return the result
